pyproject.toml
[tool.poetry]
name = "llm-in-production"
version = "0.1.0"
description = ""
authors = ["Jordi Smit <j0rd1smit@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
transformers = {extras = ["torch"], version = "^4.48.0"}
matplotlib = "^3.8.2"
tiktoken = "0.8.0"
openai = "^1.3.8"
python-dotenv = "^1.0.0"
streamlit = "^1.37.0"
scikit-learn = "^1.5.0"
langchain = "0.3.14"
faiss-cpu = "^1.7.4"
plotly = "^5.18.0"
seaborn = "^0.13.0"
sentence-transformers = "^2.2.2"
yfinance = "^0.2.35"
langchainhub = "^0.1.14"
langchain-experimental = "0.3.4"
langchain-openai = "0.3.0"
tabulate = "^0.9.0"
arxiv = "^2.1.0"
docx = "^0.2.4"
python-docx = "^1.1.2"
langchain-core = "0.3.30"
langchain-google-vertexai = "^2.0.11"
google-generativeai = "^0.8.4"
langchain-huggingface = "^0.1.2"

[tool.poetry.group.dev.dependencies]
jupyter = "^1.0.0"
invoke = "^2.2.0"
ipython = "^8.18.1"
isort = "^5.12.0"
watchdog = "^3.0.0"
pre-commit = "^3.6.0"
ipykernel = "^6.29.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.ruff]
line-length = 150
exclude = [
    'exercises',
    'streamlit/assignments',
]

[tool.black]
exclude = '''
(
  /(
    \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
    | exercises
  )/
  | foo.py           # also separately exclude a file named foo.py in
                     # the root of the project
)

[tool.isort]
profile = "black"
'''